Add git as a package
