+++ @git/test/git-test.ts
@@ -5,5 +5,5 @@ import { fetchCommits } from '../src/git.js'
 
   const range = { start: '2020-04-01', end: '2020-04-30' }
   const c = await fetchCommits(range)
-  // console.log(c)
+  console.log(c)
 })()
